requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==3.1.0
deep-translator==1.10.1
idna==3.4
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==2.0.2
